Clean up typing for persistent connection providers
